refactor locking into a common library
