+++ cli/pkg/integration/integration_test.go
@@ -2,6 +2,7 @@ package integration
 
 import (
 	"context"
+	"github.com/diggerhq/digger/libs/locking"
 	"log"
 	"math/rand"
 	"os"
@@ -14,7 +15,6 @@ import (
 	"github.com/diggerhq/digger/cli/pkg/core/terraform"
 	"github.com/diggerhq/digger/cli/pkg/digger"
 	"github.com/diggerhq/digger/cli/pkg/github/models"
-	"github.com/diggerhq/digger/cli/pkg/locking"
 	"github.com/diggerhq/digger/cli/pkg/storage"
 	"github.com/diggerhq/digger/cli/pkg/utils"
 	"github.com/diggerhq/digger/libs/comment_utils/reporting"
