+++ ee/cli/cmd/digger/main.go
@@ -10,6 +10,7 @@ import (
 	"github.com/diggerhq/digger/cli/pkg/drift"
 	comment_updater_ee "github.com/diggerhq/digger/ee/cli/pkg/comment_updater"
 	ee_drift "github.com/diggerhq/digger/ee/cli/pkg/drift"
+	core_locking "github.com/diggerhq/digger/libs/locking"
 	"log"
 	"net/http"
 	"os"
@@ -20,7 +21,6 @@ import (
 	"github.com/diggerhq/digger/cli/pkg/azure"
 	"github.com/diggerhq/digger/cli/pkg/bitbucket"
 	core_backend "github.com/diggerhq/digger/cli/pkg/core/backend"
-	core_locking "github.com/diggerhq/digger/cli/pkg/core/locking"
 	core_policy "github.com/diggerhq/digger/cli/pkg/core/policy"
 	core_storage "github.com/diggerhq/digger/cli/pkg/core/storage"
 	"github.com/diggerhq/digger/cli/pkg/digger"
